WebHacking/200-RegainSession/src/requirements.txt
bottle==0.12.13
bottle-redis==0.2.3
gevent==1.2.2
greenlet==0.4.12
PyJWT==1.5.3
redis==2.10.6
